Have some default graphDataColors
So that people actually see something when they use the
defaultGraphConfig. The colors can probably be improved.

The use of an infinite list here is justified: We are always passing it
to a `zip`-like function which will only use as many as there are data
values provided by the user.