Prefix style names, use local color names.
To better make the styles defined in the default taffybar.rc survive
reloading of gtkrc theme files (which seem to happen), this patch
prefixes all styles defined with 'taffybar-'.  It also removes the use
of gtk_color_scheme (as setting it triggers a theme reload) and defines
colors locally in the style definitions.