WorkspaceHUD: Set urgent state only on workspace containing urgent win
This fixes a bug where the urgent workspace state would be set on all
workspaces.  Now it just sets the urgent state on the workspaces
containing urgent windows.

It also reprioritizes the states so that active and visible workspaces
don't get the urgent state, even if they have urgent windows.  This
aligns with the behavior of the WorkspaceSwitcher widget and follows
from the idea that you don't need to be specially alerted to workspaces
containing urgent windows that you can already see.