Fix .IconImage spacing bug
This adds the `.Normal` class to the parent selector of `.IconImage`.
It's necessary because the icon widgets are hidden instead of removed
when the windows are closed. Without this class, the icons are hidden,
but the padding remains. Fixes #318

There should still be a default fallback icon, otherwise the padding
appears without an icon while the window is open. See #319.