markupEscapeText expects length in bytes, not characters
When text contains multi-byte characters, the length is undercalculated
for the `markupEscapeText` function, which expects the length of the
text in bytes.  This results in some labels being prematurely truncated.

I began replacing calls to `Data.Text.length` with
`Data.ByteString.length` when I read that `markupEscapeText` can also
accept the length as `-1` when the text is null-terminated, which I have
found to be the case in all the places it's used.  This approach doesn't
add a new dependency on the `bytestring` package and may be less
expensive.