windows: Fix over-escaped text in menu labels
It was observed in #379 that the window widget menu labels appear to be
wrongly escaped while the active label is OK.  This appears to be
because the GTK menuitem labels don't support markup, or at least
there seems to be no way to access the internal label to enable the
markup property.  Because of this limitation, there is no need to escape
the menuitem label text, as implemented in this change.