Refactor MPRIS2 Widget to allow customization of player widgets
This change also introduces a number of other improvements:

- It allows the wrapping of the mpris widget with an arbitary widget. This is
important because we need to set a class on the wrapping widget if we want to
hide it when no player widgets are visible. To that end we add classes
no-visible-children and visible-children to that widget so that this can be done
though css.

- This also enables #517 because it now possible (and the default behavior) that
player widgets remain visible when they are only paused and not stopped.

Fixes #310