Add dependency on `data-default`
This package is already a transitive dependency, but in order to import
directly from `Data.Default`, it must be added as a direct dependency.
The alternative is to elide the direct dependency and instead import
`Default` from one of the existing direct dependencies that re-exports
it, e.g. `import XMonad (Default)`.