flake.nix: Remove unnecessary overrides
The taffybar package in nixpkgs works without overrides and has a
maintainer (me).

So the overlay used here to build taffybar from git can be
streamlined.

The nix code for switching compiler version was removed because flake
sources only include files tracked by git (--impure or otherwise). If
we are going to edit a variable in a git-tracked file (comp.nix), then
it may as be done through editing flake.nix directly.